src/backend/requirements.txt
# ChatTrain MVP1 Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
openai>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0